+++ src/Telegram.Bot/Types/Animation.cs
@@ -16,6 +16,24 @@ namespace Telegram.Bot.Types
         public string FileId { get; set; }
 
         /// <summary>
+        /// Video width as defined by sender
+        /// </summary>
+        [JsonProperty(Required = Required.Always)]
+        public int Width { get; set; }
+
+        /// <summary>
+        /// Video height as defined by sender
+        /// </summary>
+        [JsonProperty(Required = Required.Always)]
+        public int Height { get; set; }
+
+        /// <summary>
+        /// Duration of the video in seconds as defined by sender
+        /// </summary>
+        [JsonProperty(Required = Required.Always)]
+        public int Duration { get; set; }
+
+        /// <summary>
         /// Animation thumbnail as defined by sender.
         /// </summary>
         [JsonProperty(DefaultValueHandling = DefaultValueHandling.Ignore)]