+++ test/Telegram.Bot.Tests.Integ/Admin Bot/SupergroupAdminBotTests.cs
@@ -81,7 +81,8 @@ namespace Telegram.Bot.Tests.Integ.Admin_Bot
 
             await BotClient.PinChatMessageAsync(
                 chatId: _classFixture.Chat.Id,
-                messageId: msg.MessageId
+                messageId: msg.MessageId,
+                disableNotification: true
             );
 
             _classFixture.PinnedMessage = msg;