+++ test/Telegram.Bot.Tests.Integ/Framework/TestCollectionOrderer.cs
@@ -8,13 +8,15 @@ namespace Telegram.Bot.Tests.Integ.Framework
 {
     public class TestCollectionOrderer : ITestCollectionOrderer
     {
-        private readonly string[] _orderedCollections = {
+        private readonly string[] _orderedCollections =
+        {
             // Tests that require user interaction:
             Constants.TestCollections.CallbackQuery,
             Constants.TestCollections.PrivateChatReplyMarkup,
             Constants.TestCollections.InlineMessageLiveLocation,
             Constants.TestCollections.InlineQuery,
             Constants.TestCollections.EditMessage,
+            Constants.TestCollections.EditMessageMedia,
             Constants.TestCollections.DeleteMessage,
             Constants.TestCollections.Games,
             Constants.TestCollections.Payment,
@@ -32,6 +34,7 @@ namespace Telegram.Bot.Tests.Integ.Framework
             Constants.TestCollections.SendPhotoMessage,
             Constants.TestCollections.SendVideoMessage,
             Constants.TestCollections.AlbumMessage,
+            Constants.TestCollections.ObsoleteSendMediaGroup,
             Constants.TestCollections.ReplyMarkup,
             Constants.TestCollections.LiveLocation,
             Constants.TestCollections.FileDownload,
@@ -39,6 +42,7 @@ namespace Telegram.Bot.Tests.Integ.Framework
             Constants.TestCollections.LeaveChat,
             Constants.TestCollections.GetUserProfilePhotos,
             Constants.TestCollections.EditMessage2,
+            Constants.TestCollections.EditMessageMedia2,
             Constants.TestCollections.DeleteMessage2,
             Constants.TestCollections.Stickers,
             Constants.TestCollections.Games2,
@@ -76,7 +80,7 @@ namespace Telegram.Bot.Tests.Integ.Framework
                     $"Collection \"{collection.DisplayName}\" not found in execution list.", nameof(collection));
             }
 
-            return (int)order;
+            return (int) order;
         }
     }
 }